+++ app/src/main/java/com/vultisig/wallet/data/mappers/Mapper.kt
@@ -1,9 +1,5 @@
 package com.vultisig.wallet.data.mappers
 
-interface Mapper<I, O> {
-    fun map(from: I): O
-}
-
 typealias MapperFunc<I, O> = (from: I) -> O
 
 typealias SuspendMapperFunc<I, O> = suspend (from: I) -> O
