+++ data/src/main/kotlin/com/vultisig/wallet/data/mappers/DataMappersModule.kt
@@ -18,4 +18,9 @@ internal interface DataMappersModule {
         impl: MapHexToPlainStringImpl
     ): MapHexToPlainString
 
+    @Binds
+    fun bindChainAndTokensToAddressMapper(
+        impl: ChainAndTokensToAddressMapperImpl
+    ): ChainAndTokensToAddressMapper
+
 }
