+++ data/src/main/kotlin/com/vultisig/wallet/data/mappers/Mapper.kt
@@ -1,0 +1,5 @@
+package com.vultisig.wallet.data.mappers
+
+interface Mapper<I, O> {
+    fun map(from: I): O
+}
